THE VIMENING
After giving Boon a spin, I had a longer-term realization: using a
highly idiosyncratic mode maintained by a single author (whether Xah
Fly Keys or Boon) might be subject to problems down the line, whether
due to becoming out of date or not playing nicely with other
packages (e.g. Boon not displaying the which-key popup for the C-c
shortcut). If I have to learn new keybindings to obtain composability,
I might as well learn a set of keybindings that are highly portable
and widely used (and thus likely to be maintained).